backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlmodel==0.0.21
alembic==1.13.0
psycopg2-binary==2.9.9
pydantic-settings==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.0
httpx==0.27.0
apscheduler==3.10.4
docker==7.1.0
paramiko==3.4.0
sshtunnel==0.4.0
sqlalchemy==2.0.35
PyMySQL==1.1.0
websockets==13.1
